Extracted info from weather api and display city searched
Display works regardless of whether the weather api successfully found the city or not.
Plans: add error if city not found

Managed to extract needed info from the api for the weather app.
Issue: Trying to figure out if I should make a whole new function to display everything or code in the function GetCityInfo().
Plan: Code in GetCityInfo() first then seperate later on.

Replaced Precipitation with Pressure as precipitation was not present in the JSON extracted (maybe because city was not raining). Will come back to this.